requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astronomer-cosmos[dbt-bigquery]==1.0.3
protobuf==3.20.3
apache-airflow-providers-apache-beam==5.3.0